# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NavPlat API
  description: Navigation Platform API Documentation - Complete API specification for Postman and other API clients
  version: 1.0.0
  contact:
    name: NavPlat Team
    email: support@navplat.local
servers:
  - url: http://localhost:5000/api
    description: Local Development Gateway
  - url: https://api.navplat.local
    description: Production Gateway
...
paths:
  /auth/login:
    get:
      tags: [Auth]
      summary: Initiate login flow
      description: Redirects to Keycloak for OIDC authentication with PKCE. Returns 302 redirect.
      parameters:
        - name: returnUrl
          in: query
          schema:
            type: string
          description: URL to redirect to after successful login
      responses:
        '302':
          description: Redirect to Keycloak for authentication
        '429':
          description: Too many login attempts (rate limited)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout user
      description: Revokes refresh token and clears authentication cookies
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Logout successful
        '401':
          description: Unauthorized
  /auth/user:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      description: Returns information about the currently authenticated user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
  /auth/users:
    get:
      tags: [Auth]
      summary: Get list of users
      description: Returns list of all users (for sharing journeys)
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserInfo'
        '401':
          description: Unauthorized
  /auth/refresh:
    get:
      tags: [Auth]
      summary: Refresh access token
      description: Silently refreshes the access token using the refresh token
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Token refreshed successfully
        '401':
          description: Unauthorized or refresh token expired
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: .NavPlat.Auth
      description: HTTP-only cookie containing authentication token
  schemas:
    User:
      type: object
      properties:
        userId:
          type: string
          example: "user-123"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        roles:
          type: array
          items:
            type: string
          example: ["User"]
        isAuthenticated:
          type: boolean
          example: true
    UserInfo:
      type: object
      properties:
        userId:
          type: string
          example: "user-123"
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
          example: "https://tools.ietf.org/html/rfc7231#section-6.5.1"
        title:
          type: string
          example: "Bad Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "Validation failed"
        instance:
          type: string
          format: uri
          example: "/api/journeys"
        traceId:
          type: string
          example: "0HN1234567890"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Validation errors by field name